Fix: `--type` parameter for `--invoice_list` wasn't recognized.
git-svn-id: file:///data/svn/enum/server/trunk@6575 b60c293d-5e0d-0410-87b2-fb9704ccd4b5

+++ src/cli_admin/simple.cc
@@ -423,7 +423,7 @@ main(int argc, char **argv)
     } catch (ccReg::EPP::EppError &e) {
         std::cerr << "EppError code: " << e.errCode << ", message: " 
             << e.errMsg << std::endl;
-        for (int ii = 0; ii < e.errorList.length(); ii++) {
+        for (int ii = 0; ii < (int)e.errorList.length(); ii++) {
             std::cerr << "Reason code: " << e.errorList[ii].code << ", message: " 
                 << e.errorList[ii].reason << std::endl;
         }
